fix: GenRepository 직접 의존 제거
